feat(fhirpath): strict mode, UnorderedCollection, Decimal precision, choice-type validation (#41)
* fix: workspace hygiene & critical fixes (refactor plan WS0)

- R5 crate: replace absolute foundation path with relative + adopt
  workspace inheritance (C1); codegen now emits relative-path,
  workspace-inherited manifests with tests guarding regressions
- foundation: MemoryStore/StatsRecorder use parking_lot::RwLock,
  eliminating lock-poisoning panics (C2)
- cli: delete dead ffq module that referenced nonexistent rh-ffq crate (C3)
- add [workspace.lints] (forbid unsafe_code) inherited by all members
- docs: fix WASM claim (C4), R5 status, MSRV 1.70 -> 1.91, rewrite
  ARCHITECTURE.md dependency graph to include cql/packager/R5


* test(fhirpath,fsh): add conformance harnesses and parser tests (WS2 partial)

- vendor official HL7 FHIRPath suite (937 cases) + JSON inputs; new
  harness categorizes pass/wrong/parse-err/eval-err/skipped, writes a
  machine-readable summary, and gates on a shrink-only wrong-answer
  baseline (564/935 pass at introduction)
- add rh-fhirpath CONFORMANCE.md + SPEC_COVERAGE.md recording the
  baseline and per-feature status
- fsh: 68 parser unit tests covering every entity, rule, value, and
  path form; fixes three parser bugs they exposed: URLs truncated to
  "http" in bindings/contains/from-valueset, integer quantities
  dropping their unit, single-quoted UCUM units keeping quotes
- vcl (2.10): audit claim was stale — all README symbols exist; no
  change needed

Remaining WS2 (tracked in plan): FHIRPath function waves, CQL operator
gaps, SUSHI corpus expansion.


* feat(fhirpath): UCUM duration arithmetic + equivalence + toString/toQuantity (plan 2.3 waves 11–14)

Wave 11 — UCUM duration units in Date/DateTime arithmetic:
`parse_precision_unit` extended to recognise UCUM abbreviations (`a`/`mo`/
`wk`/`d`/`h`/`min`/`s`/`ms`) alongside the calendar-duration names. The
Date-arithmetic call sites gate `'mo'` and `'a'` to an execution error
since UCUM mean-month (30.44 d) and mean-year (365.25 d) are ambiguous
over calendar boundaries; DateTime-arithmetic accepts the full set.

Wave 12 — equivalence `~` semantics per FHIRPath §6.1.5:
`values_equivalent` rewritten — strings collapse runs of whitespace and
compare case-insensitively (`'a' ~ 'A'` → true), decimals round to the
smaller fractional precision (`1.2/1.8 ~ 0.67` → true), and collections
compare as multisets (`(1|2|3) ~ (3|2|1)` → true). Falls back to strict
equality for the remaining variant pairings.

Wave 13 — toString polish:
* `Decimal.toString()` preserves the decimal point so `1.0` stringifies
  as `"1.0"`, distinguishing it from an Integer.
* `Quantity.toString()` emits calendar-duration units unquoted
  (`1 week` → `"1 week"`) while UCUM codes stay quoted (`1 'kg'`).

Wave 14 — toQuantity default unit + loose-string parser:
* `toQuantity()` on Integer/Number/Long/Boolean defaults the unit to the
  UCUM "no unit" code `'1'` per §6.3, so `1.toQuantity() = 1 '1'` works
  and `true.toQuantity()` produces `1 '1'`.
* `parse_quantity_string` now accepts unquoted calendar-duration form
  (`'1 day'`) alongside the existing UCUM-quoted form (`'5 \'mg\''`).

Five existing local unit tests updated to match the spec-correct outputs
(decimal-preserving toString, default `'1'` unit on toQuantity for
numeric and Boolean inputs).

HL7 conformance: 794 → 816 / 935 pass (84.9% → 87.3%), eval errors
41 → 34. Removed 15 stale baseline entries across testEquivalent,
testNotEquivalent, testToString, testTypes, testPlus.


* chore: untrack .claude/scheduled_tasks.lock

The lock file is session-local state for Claude Code's scheduling and
shouldn't be in the repo. Added to .gitignore and removed from tracking.


* docs: mark WS3 3.3 complete in refactor plan


* fix(fhirpath): wave 32 — unary +, $index, TypedObject, FhirPrimitive, conformsTo()

- Parser: unary + prefix now supported (+1 < +2 parses correctly)
- Evaluator: $index variable works in select()/where() iteration contexts
- Evaluator: 'is' operator returns empty for empty input (spec §6.5.4)
- Evaluator: as() errors on multi-element collections (spec §6.5.4)
- Values: TypedObject variant carries FHIR complex type name for ofType(HumanName),
  value is Age, value is Quantity, and similar type-aware operations
- Values: FhirPrimitive wrapper carries _fieldName extension data, enabling
  extension() on FHIR primitives like Patient.birthDate.extension(url)
- Metadata: apply_fhir_typing returns TypedObject for Complex field arrays
- Metadata: FhirPrimitive wraps primitives with _fieldName extension siblings
- Collection: children()/descendants()/subsetOf() handle TypedObject/FhirPrimitive
- Type ops: TypedObject checked in value_is_type/value_is_exact_type
- Type ops: fhir_is_subtype_of() supports Quantity profile hierarchy (Age, etc.)
- Functions: conformsTo(url) stub (checks base resource type)
- Functions: extension() accepts TypedString URL parameter
- Variables: %ext-{name} expands to StructureDefinition URLs
- Fixtures: codesystem-example.json gets nested sub-concepts (chol -> chol-mmol/chol-mass)
- Fixtures: observation-example.json gets patient-age extension (Age type)
- Context: with_this_value preserves index_value from outer iteration
- Context: with_this_and_index() sets both $this and $index

Result: 913 → 923/935 pass (97.6% → 98.7%)
Wrong answers: 15 → 10, parse errors: 1 → 0, eval errors: 5 → 1
Cleared: testCombine1, testExtension1, testFHIRPathAsFunction22,
         testFHIRPathIsFunction8/9, testIndex, testLiteralIntegerLessThanPolarityTrue,
         testPolymorphismIsA3, testConformsTo1/2 (conformsTo eval errors cleared)


* feat(fhirpath): strict mode, UnorderedCollection, Decimal precision, choice-type validation

- Add strict_mode and check_ordered_functions to EvaluationContext
- Introduce UnorderedCollection variant with order-dependent function guards
- Replace f64 with rust_decimal::Decimal for precision-sensitive operations
- Implement choice-type access validation ([x] suffix strict mode)
- Add FhirPrimitive/TypedObject/TypedString handling across evaluator
- Expand boundary functions with Decimal-based precision and temporal string dispatch
- Refactor arithmetic, comparison, collection, unit, and string operations
- Update conformance baselines and refactor plan progress

---------

+++ crates/rh-fhirpath/src/ast.rs
@@ -1,5 +1,6 @@
 //! Abstract Syntax Tree (AST) types for FHIRPath expressions
 
+use rust_decimal::Decimal;
 use serde::{Deserialize, Serialize};
 use std::fmt;
 
@@ -110,8 +111,8 @@ pub enum Literal {
     Boolean(bool),
     /// String literal
     String(String),
-    /// Number literal
-    Number(f64),
+    /// Decimal literal (preserves trailing zeros for precision operations)
+    Number(Decimal),
     /// Integer literal
     Integer(i64),
     /// Long number literal (explicit L suffix)
@@ -123,7 +124,10 @@ pub enum Literal {
     /// Time literal
     Time(String),
     /// Quantity literal (number with unit)
-    Quantity { value: f64, unit: Option<String> },
+    Quantity {
+        value: Decimal,
+        unit: Option<String>,
+    },
     /// DateTime precision unit literal
     DateTimePrecision(DateTimePrecision),
 }
